requirements.txt
bitsandbytes==0.43.3
scipy==1.14.0
transformers==4.44.0
peft==0.12.0
hydra-core==1.3.2
PyYAML==6.0.2
tqdm==4.66.5
av==13.1.0
sentencepiece==0.2.0
torcheval==0.0.7
tensorboard==2.18.0 
tensorboardX==2.6.2.2
accelerate==1.1.1
torch_tb_profiler==0.4.3
lightning==2.4.0